backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
pydantic==2.8.2
sqlalchemy==2.0.30
asyncpg==0.29.0
supabase==2.4.2
python-multipart==0.0.9
python-dotenv==1.0.1
alembic==1.13.1
pytest==8.2.0
httpx==0.27.0
pydantic-settings==2.0.2
clerk-backend-api
python-jose[cryptography]